app/backend/pyproject.toml
[project]
name = "gisdatamonitor-backend"
version = "0.1.0"
description = "GISDataMonitor Jiangxi MVP backend"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10.4",
  "fastapi>=0.116.0",
  "feedparser>=6.0.11",
  "geopandas>=1.0.1",
  "httpx>=0.28.1",
  "mapbox-vector-tile>=2.2.0",
  "mercantile>=1.2.1",
  "numpy>=2.2.0",
  "orjson>=3.11.0",
  "pydantic>=2.11.0",
  "pydantic-settings>=2.10.0",
  "python-dateutil>=2.9.0",
  "python-dotenv>=1.1.0",
  "rasterio>=1.4.0",
  "requests>=2.32.4",
  "shapely>=2.1.0",
  "sqlalchemy>=2.0.38",
  "uvicorn>=0.35.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.4.1",
]

postgres = [
  "psycopg[binary]>=3.2.9",
]

[build-system]
requires = ["setuptools>=70.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
